requirements.txt
certifi==2018.4.16
chardet==3.0.4
idna==2.7
mysql-connector-python==8.0.11
protobuf==3.6.0
requests==2.19.1
six==1.11.0
urllib3==1.23
wget==3.2
